Improved documentation and Trajectory sampling functionality.
Documentation and style improvements mostly in the Trajectory class, but
also some in the Reaction class.

Improved sampling functionality in Trajectory: Renamed the sample_linspace
function to sample_state_seq and changed it to take an arbitrary sorted
sequence (easier for user to specify linspace parameters that way). Also
added function to sample single state using bisection, intended for use
with delayed reactions.