Reworked Reactant class - now supports higher-order reactions.
Reactant class now uses different input specifications for reactants and
products. The reactants list is now in the same format as the state
vector; this allows easier computation (and more logical
specification) of the reaction propensity in the case of higher-order
reactions. Also added a convenience method method to obtain the reverse
of a reaction.