Added a convenience function to return Ensemble prob. dist.
The function returns the probability distribution estimated by the
ensemble. Repercussions include constraining bin indices to be contiguous
and zero-based from now on - this shouldn't be too restrictive, as the
Paving class can always be extended to accept queries for an index and
return more information about the bin.

Conflicts:
	python/ensemble.py