Fixed critical bug in trajectory cloning.
The initial time of the clones was being set to the current time, which
caused the past-state sampling to behave in an unexpected way (using the
initial state for sampling earlier than the trajectory was spawned).

The cloning method now sets the new trajectories' initial times to that
of the parent trajectory, making the past-state sampling behave
correctly. The fix has been tested and found to work.